refactor(core): move report header, redaction, and zip export to tyutool-core
Continue the log-tooling consolidation (P1b): report-info header, credential
redaction, and zip export move from src-tauri/src/logs.rs into
tyutool_core::diagnostics, gated behind a new zip Cargo feature so
tyutool-serve and tyutool-bridge do not link the zip crate.

build_report_info / REDACT_PREFIXES / redact_log_content stay unconditional
(no zip dependency); write_logs_zip / gather_and_write_logs_zip require the
zip feature. src-tauri now enables tyutool-core features = [libudev, zip]
and drops its own direct zip dependency, since all zip:: usage (including the
zip-verification tests) moved with the code.

export_logs_zip (Tauri command) and the archive_batch_cmd call site now call
tyutool_core::gather_and_write_logs_zip directly, no forwarding wrapper.

Splitting the tiers this way opened a coverage hole that had to be closed in
the same change: redaction became callable without the zip feature, while
every test that exercised it sat behind that feature, so the default
cargo test -p tyutool-core covered the mask contract zero times. AGENTS.md
treats that contract as a security guarantee, so four direct tests now live in
the unconditional module: each of the four prefixes is masked, values of 12,
16 and 20+ characters are all masked (device identifiers are not fixed-length),
non-credential content is returned byte-identical, and, asserting observed
rather than assumed behavior, a prefix embedded mid-token still triggers
redaction. That last one over-redacts, which is the safe direction for this
contract; it is recorded rather than changed, because this stage moves code
and does not alter behavior.
