+++ crates/tyutool-core/Cargo.toml
@@ -8,6 +8,12 @@ description = "Shared flash/erase plugin registry and serial helpers for tyutool
 # Enables libudev-backed port enumeration on Linux (GUI / glibc builds only).
 # CLI musl builds and CI must leave this disabled; the Tauri shell enables it.
 libudev = ["serialport/libudev"]
+# Enables zip-based log export/archive bundling (write_logs_zip /
+# gather_and_write_logs_zip). The GUI must enable it (export-for-report and
+# archive-batch both zip session logs); tyutool-serve and tyutool-bridge must
+# leave it disabled so they don't link the zip crate for a capability they
+# don't use.
+zip = ["dep:zip"]
 
 [dev-dependencies]
 serde_json = "1"
@@ -23,3 +29,4 @@ crc32fast = "1"
 log = "0.4"
 espflash = { version = "4.3", default-features = false, features = ["serialport"] }
 md-5 = "0.10"
+zip = { version = "2", default-features = false, features = ["deflate"], optional = true }
